pyproject.toml
[tool.poetry]
name = "labsrestful"
version = "0.1.0"
description = ""
authors = ["michkail <michkail@sateknologi.id>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "4"
python-dotenv = "^1.0.0"
psycopg-binary = "^3.1.12"
django-cassandra-engine = "^1.8.0"
pillow = "^10.2.0"
qrcode = "^7.4.2"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
